% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{correlation_matrix}
\alias{correlation_matrix}
\title{Return a correlation matrix}
\usage{
correlation_matrix(year, data = NBA)
}
\arguments{
\item{year}{The year that you are interested in.}
}
\description{
This function allows you obtain a correlation martix of all numeric variables in the NBA dataset for a given year
}
\examples{
correlation_matrix(2000)
}
\keyword{NBA,}
\keyword{year}
