% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_year.R
\name{filter_year}
\alias{filter_year}
\title{Filter NBA dataset by year}
\usage{
filter_year(year, data = NBA)
}
\arguments{
\item{year}{The year which you are interested in. No default}
}
\description{
This function allows you to filter the NBA dataset by year.
}
\examples{
filter_year(2000)
}
\keyword{NBA,}
\keyword{year}
