% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA}
\alias{NBA}
\title{Summary statistics of NBA seasons.}
\format{
A data frame with 24691 rows and 52 variables:
\describe{
\item{Year}{NBA year}
\item{Player}{Name of players}
...
}
}
\source{
\url{https://www.kaggle.com/drgilermo/nba-players-stats/data}
}
\usage{
NBA
}
\description{
A dataset containing the information about NBA players
}
\keyword{datasets}
