% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player_year.R
\name{oldest_player_year}
\alias{oldest_player_year}
\title{Select the oldest NBA player of a year}
\usage{
oldest_player_year(year, data = NBA)
}
\arguments{
\item{year}{The year that you are interested in.}
}
\description{
This function allows you know who is the oldest NBA player of a year
}
\examples{
oldest_player_year(2000)
}
\keyword{NBA,}
\keyword{year}
